Set up tooling: ruff + pytest config, CI workflow, lint/format pass
- pyproject.toml with ruff (py312, E/F/W/I/UP/B/SIM/C4/RUF) and pytest config
- fix all lint findings, apply ruff format
- .github/workflows/ci.yml: lint+test, hassfest, HACS validation
- manifest.json: alphabetical key order for hassfest

All 29 tests pass; ruff check + format clean.

+++ custom_components/tuya_lan/helpers.py
@@ -46,4 +46,4 @@ def scale_to_device(value: float, options: dict[str, Any]) -> int | float:
     scale = options.get("scale", 1) or 1
     offset = options.get("offset", 0) or 0
     raw = (value - offset) * scale
-    return int(round(raw)) if options.get("integer", True) else raw
+    return round(raw) if options.get("integer", True) else raw
